Safari-611.1.10/LayoutTests/fast/events/ios/no-touch-events-when-stopping-momentum-scroll-in-overflow-expected.txt

PASS document.getElementById('container').scrollTop became different from 0
PASS document.getElementById('touches').textContent is "1"
PASS document.getElementById('touches').textContent is "1"
PASS document.getElementById('touches').textContent is "2"
PASS successfullyParsed is true

TEST COMPLETE
To manually test, swipe down to scroll in the container. The touch count should update to 1. Tap to stop scrolling while decelerating. The touch count should remain at 1. Then tap again in the container. The touch count should be 2.

The touch count is: 2
